requirements.txt
Flask>=2.2
gunicorn>=20.1
python-dotenv>=1.0
google-api-python-client>=2.0
google-auth-httplib2>=0.1
google-auth-oauthlib>=1.2
google-auth>=2.0
httplib2>=0.20
requests>=2.0
oauthlib>=3.0
Werkzeug>=2.0
